Server CI: fix flaky CI due to PG not being ready
I think that what was going on is that PG was not yet ready when we were
trying to run the migrations. We now use the wait-for fix in #554 to make
it properly wait for PG to be ready. Under normal circumstances the wait
should be zero, so it probably won't affect CI speed.